+++ client/src/main/java/com/influxdb/client/TasksApi.java
@@ -25,6 +25,7 @@ import java.time.OffsetDateTime;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.annotation.concurrent.ThreadSafe;
 
 import com.influxdb.client.domain.Label;
 import com.influxdb.client.domain.LabelResponse;
@@ -44,6 +45,7 @@ import com.influxdb.client.domain.User;
  *
  * @author Jakub Bednar (bednar@github) (11/09/2018 07:54)
  */
+@ThreadSafe
 public interface TasksApi {
 
     /**